backend/requirements.txt
Django==4.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1
openpyxl==3.1.5
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0 